\name{IsoDeconvMM-package}
\alias{IsoDeconvMM-package}
\alias{IsoDeconvMM}
\docType{package}
\title{
\packageTitle{IsoDeconvMM}
}
\description{
\packageDescription{IsoDeconvMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{IsoDeconvMM}
\packageIndices{IsoDeconvMM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{IsoDeconvMM}

Maintainer: \packageMaintainer{IsoDeconvMM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
