% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoDeconvMM.R
\name{isoDeconvMM}
\alias{isoDeconvMM}
\title{Title
 
\code{isodeconvMM} ... description}
\usage{
isoDeconvMM(directory = NULL, mix_files, pure_ref_files, fraglens_files,
  bedfile, knownIsoforms, discrim_genes, readLen, lmax = 600,
  eLenMin = 1, mix_names = NULL, initPts = NULL,
  sim = simControl())
}
\arguments{
\item{directory}{an optional character string denoting the path to the directory where all of the 
mix_files, pure_ref_files, fraglens_files, and bedfile are located.}

\item{mix_files}{a vector of the file names for the count files (in .txt formats) 
for the samples containing mixtures of cells (add more details)}

\item{pure_ref_files}{a data.frame or matrix where the first column is the file names for the 
count files (in .txt formats) for the pure reference cell type samples and the second column
contains the names of the pure cell type associated with each sample}

\item{fraglens_files}{a vector of the file names for the fragment length files. This data
should contain two columns, one for () and another for ()}

\item{bedfile}{name of the .bed file associated with the data}

\item{knownIsoforms}{character string for the name of an RData object. This RData object is a
list where each element of the list corresponds to a transcript cluster
and contains a matrix of 0s and 1s, where the rows correspond to exons and the columns correspond 
to isoforms. Instructions for creating such an RData object can be found in the (isoforms vignette)}

\item{discrim_genes}{vector of genes that are suspected to have differential gene expression 
(perhaps based on CuffLinks output)}

\item{readLen}{numeric value of the read length of the RNAseq experiment}

\item{lmax}{numeric value of the maximum fragment length of the experiment}

\item{eLenMin}{numeric value of the minimum value of effective length which allows for errors 
in the sequencing/mapping process}

\item{mix_names}{an optional vector of the desired nicknames of the mixture samples corresponding,
in the same order, to the mix_files list. If left as the defaul NULL value, the nicknames used
will be the names given in the mix_files minus the .txt extension}

\item{initPts}{an optional matrix of initial probability estimates for the cell composition
of the mixture samples to be used in the optimization procedure. The matrix should have k columns,
where k = number of pure cell types of interest. Each row corresponds to different combinations of 
initial probability values. The column names of the matrix must correspond to the pure cell type
names given in the second column of the pure_ref_files object (no particular ordering needed)}
}
\value{
A list object with the following elements (...)
}
\description{
Title
 
\code{isodeconvMM} ... description
}
